Integ-tests: create mount targets when creating EFS
After https://github.com/aws/aws-parallelcluster/pull/4532, ParallelCluster no long creates mount targets for external EFS file systems. ParallelCluster CLI validates the external EFS file systems have necessary mount targets. Therefore, we need to change our integration tests to create mount targets for any external EFS file systems.
