+++ tests/integration-tests/tests/update/test_update.py
@@ -777,6 +777,7 @@ def test_dynamic_file_systems_update(
     request,
     snapshots_factory,
     efs_stack_factory,
+    mount_target_stack_factory,
     vpc_stack,
     key_name,
     s3_bucket_factory,
@@ -815,6 +816,7 @@ def test_dynamic_file_systems_update(
         vpc_stack,
         region,
         efs_stack_factory,
+        mount_target_stack_factory,
         fsx_factory,
         svm_factory,
         open_zfs_volume_factory,
@@ -947,6 +949,7 @@ def _create_shared_storages_resources(
     vpc_stack,
     region,
     efs_stack_factory,
+    mount_target_stack_factory,
     fsx_factory,
     svm_factory,
     open_zfs_volume_factory,
@@ -957,7 +960,9 @@ def _create_shared_storages_resources(
     ebs_volume_id = snapshots_factory.create_existing_volume(request, vpc_stack.cfn_outputs["PublicSubnetId"], region)
 
     # create 1 efs
-    existing_efs_id = efs_stack_factory(1)[0]
+    existing_efs_ids = efs_stack_factory(1)
+    mount_target_stack_factory(existing_efs_ids)
+    existing_efs_id = existing_efs_ids[0]
 
     # create 1 fsx lustre
     import_path = "s3://{0}".format(bucket_name)
